pyproject.toml
[project]
name = "ollamabot"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "ollama>=0.3.3",
    "click>=8.1.7",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ollamabot"]
